generate 구 현

See merge request s08-final/S08P31B106!211

+++ user-service/src/main/java/com/gemini/userservice/dto/request/RequestGeneratePoseDto.java
@@ -15,5 +15,5 @@ public class RequestGeneratePoseDto {
 
     private List<String> users;
 
-    private Long sample;
+    private Integer sample;
 }
